Philo is working
Finally, I manage to solve the problem with the threads priorities. It
was related to the pthread_create function and the monitor thread. I was
passing the adress of the pointer to the structure table, instead the variable itself.

I solved other problems as well, related with some of the mutex
functions, which didn't unlock the mutexes.

I think the program is working, but I want to spent some time upgrading
the error handling.